+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add support for Odyssey's on-foot bindings
+- Add hardcoded hotkeys (normal/hi-res screenshot, toggle FPS/HUD/bandwidth display)
 
 ## [1.13.1](https://github.com/poveden/EliteChroma/compare/v1.13.0...v1.13.1) — 2021-06-22